rename 'abis/resolver' to 'abis/shared'
