+++ packages/datasources/src/index.ts
@@ -3,4 +3,4 @@ export * from "./namespaces";
 
 // export shared ABIs for consumer convenience
 export { ResolverABI } from "./lib/resolver";
-export { StandaloneReverseRegistrar as StandaloneReverseRegistrarABI } from "./abis/resolver/StandaloneReverseRegistrar";
+export { StandaloneReverseRegistrar as StandaloneReverseRegistrarABI } from "./abis/shared/StandaloneReverseRegistrar";
